Add files via upload
Parts list for what we ordered, we got the vast majority of parts from jagr electronics on the 18th of November.  We still need micro SD cards and are trying an alternative to max 667 circuits. We have pushed our circuit prototyping to the 21st when we build the circuits. The overdischarge protection circuit may be pushed back if our cheaper solution does not work, but is not a major issue since the over-discharge circuit is not essential to the build. We can get rid of the ODPC if we really need to.