Add files via upload
These pdf files are datasheets for the screen we used, after some reading, we figured out that our voltage is fine and in theory, we will not need to remove any IC's like in the project.  To get our emulator to run I had to first format our micro sd, then we flashed it with an image file containing the emulator. Once that was on their we first put in into a PI 1 and connected it to wifi where I connected to it from my laptop and logged in as a root user and edited the config files. Once the config files were done we went back to soldering  and we got it to work. I've included a picture of our test.