Add files via upload
After many issues with screen breaking and out project falling apart in general, something has been completed. Today the pi was transplanted into a double sided plastic case along with the batteries required to power the pi. Since the screens we had broke a backup plan a system to play retro games using an hdmi to the screen without the need of power cables or controllers. I have also made the button system which will be installed in the reverse side of the case from the battery pack. The device is now powered by 4 AA batteries and one 9mm socket.